+++ pdf/test/arabic_test.dart
@@ -439,7 +439,7 @@ void main() {
         65176,
         65252,
         65247,
-        1575
+        1575,
       ]),
     ];
 
@@ -484,6 +484,38 @@ void main() {
                   'القهوة مشروب يعد من بذور الب المحمصة، وينمو في أكثر من 70 لداً. خصوصاً في المناطق الاستوائية في أمريكا الشمالية والجنوبية وجنوب شرق آسيا وشبه القارة الهندية وأفريقيا. ويقال أن البن الأخضر هو ثاني أكثر السلع تداولاً في العالم بعد النفط الخام.',
               style: TextStyle(
                 fontSize: 20,
+              ),
+            ),
+          ],
+        ),
+      ),
+    ));
+  });
+
+  test(
+      'Text Widgets, Mixed Arabic and Latin words should be rendered in order ',
+      () {
+    pdf.addPage(Page(
+      textDirection: TextDirection.rtl,
+      build: (Context context) => RichText(
+        text: TextSpan(
+          text: 'النصوص ثنائية الإتجاه Bidirectional Text\n',
+          style: TextStyle(
+            font: arabicFont,
+            fontSize: 30,
+          ),
+          children: const <TextSpan>[
+            TextSpan(
+              text: r'''
+الكلمات اللاتينية المضافة إلى نص عربي يجب أن توضع في الترتيب الصحيح Right Order مهما كان موضعها في النص.
+At the Beginning of the sentence في بداية الجملة
+أو في منتصفها In the middle of the sentence حيث يكون بعدها كلام عربي
+أو في نهاية النص At the end of the sentence
+أيضا ترتيب الأرقام والرموز يجب 1 أن 2 يكون 3 صحيحاً$.
+ولا ننسى أيضا فواصل السطور Line breakers حيث وجودها في موضعها الصحيح مهم جدا في النصوص ثنائية الاتجاه Bidirectional
+              ''',
+              style: TextStyle(
+                fontSize: 18,
               ),
             ),
           ],